[arch]: merge 2 files into one
